// src/app/core/enums/common.enum.ts
export enum ViewMode{
  Simple = "simple",
  Advanced = "advanced",
}

export enum ViewAngle{
  Front = "front",
  Back = "back",
  //Lateral = "lateral",
}

export enum Side{
  Left = "left",
  Right = "right",
}

export enum PressureLevel {
  Low = 'low',
  Medium = 'medium',
  High = 'high',
  VeryHigh = 'very high',
}

export enum AlertStatus {
  None = 'none',
  Warning = 'warning',
  Critical = 'critical',
}
export enum EquipmentPrivacy {
  All = 'all',
  Created = 'created',
}